rag/requirements.txt
# RAG Service Dependencies
flask==3.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0

# CPU-only PyTorch (install before sentence-transformers to avoid GPU version)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchvision==0.16.0+cpu

sentence-transformers==2.3.1
PyPDF2==3.0.1
pgvector==0.2.4
requests==2.31.0
werkzeug==3.0.1
kafka-python==2.0.2

